pyproject.toml
[project]
name = "multi-agent-system-langraph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=6.29.5",
    "ipython>=9.3.0",
    "langchain>=0.3.25",
    "langchain-groq>=0.3.2",
    "langgraph>=0.4.8",
    "langsmith>=0.3.45",
    "python-dotenv>=1.1.0",
]
